snv-git: no tags on case insensitive file systems
When running on a case-insenstive file system,
checking out a version may end up in a problem:

When both "File" and "FILE" are present, the worktree seems to be
dirty to Git, so that we can not checkout another version.

Disable the branch -> tag conversion for case insensitive files systems,
until we find a better solution.
(The better solution may be to check out a temp branch, commit, checkout
another SVN branch and delete the temp branch)